requirements.txt
flask
flask-sqlalchemy
flask-cors
requests
better_profanity
python-dotenv